+++ Spy/Common/Formatters/SpyDateFormatter.swift
@@ -1,0 +1,17 @@
+//
+//  SpyDateFormatter.swift
+//  Spy
+//
+//  Created by Tomasz Lewandowski on 07/02/2020.
+//  Copyright © 2020 AppUnite Sp. z o.o. All rights reserved.
+//
+
+import Foundation
+
+public enum SpyDateFormatter {
+    public static var `default`: DateFormatter {
+        let formatter = DateFormatter()
+        formatter.dateFormat = "YYYY-MM-dd HH:mm:ss.SSSSSSZ"
+        return formatter
+    }
+}