+++ Spy/Common/Spies/FileSpy/FileManager+PFileManager.swift
@@ -1,0 +1,13 @@
+//
+//  FileManager+PFileManager.swift
+//  Spy
+//
+//  Created by Tomasz Lewandowski on 06/02/2020.
+//  Copyright © 2020 AppUnite Sp. z o.o. All rights reserved.
+//
+
+import Foundation
+
+extension FileManager: PFileManager {
+    
+}