+++ Spy/Common/Spies/ThreadSafeSpy/NSLock+PLock.swift
@@ -1,0 +1,13 @@
+//
+//  NSLock+PLock.swift
+//  Spy
+//
+//  Created by Tomasz Lewandowski on 07/02/2020.
+//  Copyright © 2020 AppUnite Sp. z o.o. All rights reserved.
+//
+
+import Foundation
+
+extension NSLock: PLock {
+    
+}