+++ CHANGELOG.md
@@ -6,6 +6,13 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 ## Unreleased
 
+## [0.5.0]
+- Use DateFormatter to format log timestamp
+- Change time format to 'YYYY-MM-dd HH:mm:ss.SSSSSSZ'
+- Introduce FileSpy that allows to log into files
+- Remove NetworkSpy from example & readme
+- Added support for thread safe spies
+
 ## [0.4.0]
 - Added Spied property wrapper for logging property access & changes